alterando coluna de telefone para pg

+++ src/database/migrations/00_create_user.js
@@ -5,7 +5,7 @@ exports.up = async function (knex) {
     table.string('cpf', 11).notNullable();
     table.string('email').notNullable().unique();
     table.string('password', 8).notNullable();
-    table.integer('phone').notNullable();
+    table.bigInteger('phone').notNullable();
     table.boolean('is_admin').notNullable();
   })
 }
